* commit '2612fc99024dc13d3b92fa9239e92c9be1e130aa':
  Change names of toplevel and file name
  Change how the source package dir is named
  Add WIN32 cpack generator
  reduce warnings
  Remove failing example from list
  Add exit() value
  Warning reduction

+++ hdf/examples/GR/h4ex_GR_create_and_write_image.c
@@ -56,8 +56,8 @@ int main( )
    {
       for (j = 0; j < X_LENGTH; j++)
       {
-         image_buf[i][j][0] = (i + j) + 1;     /* first component */
-         image_buf[i][j][1] = (i + j) + 1;     /* second component */
+         image_buf[i][j][0] = (int16)(i + j) + 1;     /* first component */
+         image_buf[i][j][1] = (int16)(i + j) + 1;     /* second component */
       }
     }
